Current working version, before further updates.
Library has a working srsTest() where registers are access through current configuration
files.  Result has the return from the FEC to standard out.

Current config files use 10.0.1.2, which is the test FEC through the test SRU.
Modified the rest of the config files so that they're a bit easier to read.